statadd/statclear API calls
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@10277 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ lib/rts2/camd.cpp
@@ -970,7 +970,7 @@ int Camera::initValues ()
 #ifdef __linux__
 			// read limit
 			std::ifstream ifs ("/proc/sys/kernel/shmmax");
-			size_t shmax;
+			long long shmax;
 			ifs >> shmax;
 			ifs.close ();
 			if (ifs.fail ())
